auth list planos with join and planAss update legacy fn done

+++ src/controllers/plano-assinatura.controller.js
@@ -1,13 +1,15 @@
 const id6 = require("../helpers/id6");
 const mongoose = require("mongoose");
-const PlanAssSchema = require("../schemas/plano-assinatura.schema");
 const {
   TIPO_PLANO_ASSINATURA,
   TIPO_FEATURE_PLANO_ASSINATURA,
   TIPO_VALOR_FEATURE_PLAN_ASS,
 } = require("../schemas/enums");
+const PlanAssSchema = require("../schemas/plano-assinatura.schema");
+const UsuarioSchema = require("../schemas/usuario.schema");
 
 const PlanAssModel = mongoose.model("PlanoAssinatura", PlanAssSchema);
+const UsuarioModel = mongoose.model("Usuario", UsuarioSchema);
 
 const tiposPlanAss = Object.values(TIPO_PLANO_ASSINATURA);
 const tiposFeaturePlanAss = Object.values(TIPO_FEATURE_PLANO_ASSINATURA);
@@ -19,28 +21,30 @@ const TIPO_FEATURE_VALOR = (() => {
   const F = TIPO_FEATURE_PLANO_ASSINATURA;
   const V = TIPO_VALOR_FEATURE_PLAN_ASS;
   return [
-    { tipo: T.PF, feat: F.VER_DASHBOARD, val: V.VER },
-    { tipo: T.PF, feat: F.VER_G_VAGAS_MES, val: V.VER },
-    { tipo: T.PF, feat: F.VER_G_VAGAS_ESTADO, val: V.VER },
-    { tipo: T.PF, feat: F.VER_G_CONTRATACOES_MES, val: V.VER },
-    { tipo: T.PF, feat: F.VER_G_COMPETENCIAS_ESTADO, val: V.VER },
-    { tipo: T.PF, feat: F.VER_NOME_EMPRESA, val: V.VER },
-    { tipo: T.PF, feat: F.LIMITE_CANDIDATURAS, val: V.LIMITE },
+    { tipo: T.PF, chave: F.VER_DASHBOARD, valType: V.VER },
+    { tipo: T.PF, chave: F.VER_G_VAGAS_MES, valType: V.VER, parent: F.VER_DASHBOARD },
+    { tipo: T.PF, chave: F.VER_G_VAGAS_ESTADO, valType: V.VER, parent: F.VER_DASHBOARD },
+    { tipo: T.PF, chave: F.VER_G_CONT_MES, valType: V.VER, parent: F.VER_DASHBOARD },
+    { tipo: T.PF, chave: F.VER_G_COMP_ESTADO, valType: V.VER, parent: F.VER_DASHBOARD },
+    { tipo: T.PF, chave: F.VER_NOME_EMPRESA, valType: V.VER },
+    { tipo: T.PF, chave: F.LIMITE_CANDIDATURAS, valType: V.LIMITE },
 
-    { tipo: T.PJ, feat: F.VER_DASHBOARD, val: V.VER },
-    { tipo: T.PJ, feat: F.VER_G_VAGAS_MES, val: V.VER },
-    { tipo: T.PJ, feat: F.VER_G_VAGAS_ESTADO, val: V.VER },
-    { tipo: T.PJ, feat: F.VER_G_CONTRATACOES_MES, val: V.VER },
-    { tipo: T.PJ, feat: F.VER_G_COMPETENCIAS_ESTADO, val: V.VER },
-    { tipo: T.PJ, feat: F.VER_DADOS_CANDIDATO, val: V.VER },
-    { tipo: T.PJ, feat: F.VER_CV_FULL, val: V.VER },
-    { tipo: T.PJ, feat: F.LIMITE_VAGAS, val: V.LIMITE },
+    { tipo: T.PJ, chave: F.VER_DASHBOARD, valType: V.VER },
+    { tipo: T.PJ, chave: F.VER_G_VAGAS_MES, valType: V.VER, parent: F.VER_DASHBOARD },
+    { tipo: T.PJ, chave: F.VER_G_VAGAS_ESTADO, valType: V.VER, parent: F.VER_DASHBOARD },
+    { tipo: T.PJ, chave: F.VER_G_CONT_MES, valType: V.VER, parent: F.VER_DASHBOARD },
+    { tipo: T.PJ, chave: F.VER_G_COMP_ESTADO, valType: V.VER, parent: F.VER_DASHBOARD },
+    { tipo: T.PJ, chave: F.VER_DADOS_CANDIDATO, valType: V.VER },
+    { tipo: T.PJ, chave: F.VER_CV_FULL, valType: V.VER },
+    { tipo: T.PJ, chave: F.LIMITE_VAGAS, valType: V.LIMITE },
+
+    { tipo: T.MA, chave: F.ADMIN, valType: V.VER },
   ];
 })();
 
 const tipoFeatureValorCheck = (tipoValor, valorFeature) => {
   if (tipoValor === TIPO_VALOR_FEATURE_PLAN_ASS.VER) {
-    if (typeof valorFeature === "boolean") {
+    if (valorFeature == 1) {
       return true;
     }
   }
@@ -57,13 +61,13 @@ const tipoFeatureValorCheck = (tipoValor, valorFeature) => {
 // id "1,2,3"
 // tipo "PF"|"PJ"|"MA"
 // active "true"|"false"
-async function listPlanosAssinatura(paramSearch) {
+async function listPlanosAssinatura(paramSearch = {}) {
   const search = {};
   if (paramSearch.id) {
     if (paramSearch.id instanceof Array) {
-      search._id = { $in: paramSearch.id };
+      search._id = { $in: paramSearch._id };
     } else {
-      search._id = paramSearch.id;
+      search._id = paramSearch._id;
     }
   }
   if (paramSearch.tipo) {
@@ -77,12 +81,19 @@ async function listPlanosAssinatura(paramSearch) {
   let planAssData = await PlanAssModel.find(search).exec();
 
   return {
-    data: planAssData,
+    data: (planAssData || []),
     meta: {
       total,
       search,
     },
   };
+}
+async function listTiposPlanosAssinatura() {
+  return [
+    { tipo: TIPO_PLANO_ASSINATURA.PF, nome: "PF - Candidato" },
+    { tipo: TIPO_PLANO_ASSINATURA.PJ, nome: "PJ - Empresa" },
+    { tipo: TIPO_PLANO_ASSINATURA.MA, nome: "Master Admin" },
+  ]
 }
 async function showDefaultPlanoAssinatura(tipo) {
   if (!tiposPlanAss.includes(tipo)) {
@@ -110,10 +121,11 @@ async function makePlanosAssinaturaNotDefault(tipo) {
     { new: true, runValidators: true }
   );
 
-  console.info(`\nAtualizou ${undefaultedPlanos.length} para não padrão\n`);
+  console.info(`\nAtualizou ${undefaultedPlanos?.length} para não padrão\n`);
 }
 async function showPlanoAssinatura(id) {
-  return await PlanAssModel.findById(id);
+  const planoAssinatura = await PlanAssModel.findById(id);
+  return planoAssinatura
 }
 async function listFeatures(tipo) {
   if (tipo) {
@@ -142,9 +154,8 @@ async function savePlanoAssinatura(paramData) {
       );
     }
   }
-  if (paramData.active) {
-    planAssData.active = !!paramData.active;
-  }
+  planAssData.active = !!paramData.active;
+
   if (paramData.preco != null && !isNaN(paramData.preco)) {
     planAssData.preco = +paramData.preco;
   }
@@ -159,15 +170,15 @@ async function savePlanoAssinatura(paramData) {
       const tipoFeatureValor = TIPO_FEATURE_VALOR.find((tfv) => {
         return (
           tfv.tipo === paramData.tipo &&
-          tfv.feat === feature.key &&
-          tipoFeatureValorCheck(tfv.val, feature.value)
+          tfv.chave === feature.chave &&
+          tipoFeatureValorCheck(tfv.valType, feature.valor)
         );
       });
 
       if (tipoFeatureValor) {
         planAssFeatures.push({
-          key: feature.key,
-          value: feature.value,
+          chave: feature.chave,
+          valor: +feature.valor,
         });
       }
     });
@@ -188,8 +199,8 @@ async function savePlanoAssinatura(paramData) {
     planAssData.defaultForTipo = true;
   }
 
-  if (paramData.id) {
-    planAssData._id = paramData.id;
+  if (paramData._id) {
+    planAssData._id = paramData._id;
 
     if (!paramData.defaultForTipo) {
       const otherDefaultPlanAss = await PlanAssModel.countDocuments({
@@ -200,27 +211,31 @@ async function savePlanoAssinatura(paramData) {
       if (otherDefaultPlanAss < 1) {
         throw new Error(
           [
-            `Falha ao atualizar plano de assinatura: Nenhum outro plano padrão`,
-            `para o tipo "${planAssData.tipo}" encontrado, e este plano seria`,
-            `salvo como não-padrão - é obrigatório pelo menos um plano padrão para cada tipo.`,
+            `Falha ao atualizar plano de assinatura para o tipo "${planAssData.tipo}":`,
+            `é obrigatório 1 plano padrão para cada tipo.`,
           ].join(" ")
         );
       }
     }
 
-    return await PlanAssModel.findByIdAndUpdate(paramData.id, planAssData, {
+    return await PlanAssModel.findByIdAndUpdate(paramData._id, planAssData, {
       new: true,
       runValidators: true,
     });
   } else {
     planAssData._id = id6();
-    planAssData.ownerId = paramData._id;
     return await PlanAssModel.create(planAssData);
   }
 }
 async function deletePlanoAssinatura(id) {
   const planAss = await PlanAssModel.findById(id);
   if (!planAss) throw new Error("Plano de Assinatura não encontrado");
+
+  if (planAss.defaultForTipo) {
+    throw new Error(`Operação não permitida: deletar plano padrão para tipo "${planAss.tipo}"`);
+  }
+
+  // check if users have the plan
 
   await PlanAssModel.findByIdAndDelete(id);
 
@@ -233,8 +248,22 @@ async function deletePlanoAssinatura(id) {
 // migration
 
 async function legacyUpdateUsers() {
-  // get all users
-  // update legacy plano for each
+  const defaultPF = await showDefaultPlanoAssinatura(TIPO_PLANO_ASSINATURA.PF)
+  const defaultPJ = await showDefaultPlanoAssinatura(TIPO_PLANO_ASSINATURA.PJ)
+  
+  await UsuarioModel.updateMany({ plano: defaultPF._id }, { plano: { $in: [
+    "PF_FREE",
+    "PF_SMART",
+    "PF_PREMIUM",
+  ] }})
+  await UsuarioModel.updateMany({ plano: defaultPJ._id }, { plano: { $in: [
+    "PJ_FREE",
+    "PJ_SMART",
+    "PJ_PREMIUM",
+    "PJ_ENTERPRISE",
+  ] }})
+
+  return true;
 }
 
 // router fns
@@ -243,7 +272,7 @@ async function handleGet(req, res) {
   const search = {};
 
   if (req.query.id) {
-    search.id = req.query.id.split(",");
+    search._id = req.query.id.split(",");
   }
   if (req.query.tipo) {
     search.tipo = req.query.tipo;
@@ -272,6 +301,9 @@ async function handleGet(req, res) {
 async function handleGetFeatures(req, res) {
   return listFeatures(req.query.tipo);
 }
+async function handleGetTipos(req, res) {
+  return listTiposPlanosAssinatura();
+}
 async function handleGetSingle(req, res) {
   return await showPlanoAssinatura(req.params.id);
 }
@@ -288,6 +320,7 @@ module.exports = {
 
   showDefaultPlanoAssinatura,
   listPlanosAssinatura,
+  listTiposPlanosAssinatura,
   showPlanoAssinatura,
   listFeatures,
   savePlanoAssinatura,
@@ -297,6 +330,7 @@ module.exports = {
 
   handleGet,
   handleGetFeatures,
+  handleGetTipos,
   handleGetSingle,
   handleDelete,
   handlePost,
